% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capitalize.R
\name{Capitalize}
\alias{Capitalize}
\title{Capitalize}
\usage{
Capitalize(string)
}
\arguments{
\item{dataset}{Dataset to investigate}
}
\description{
This function capitalizes the first letter of a normal string
}
\examples{
Capitalize()
}
\keyword{capitalize}
\keyword{string,}
