% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountNum.R
\name{CountNum}
\alias{CountNum}
\title{CountNum}
\usage{
CountNum(x, na.rm)
}
\arguments{
\item{dataset}{Dataset to investigate}
}
\description{
This function counts the number with same syntzx as mean, max, min.
}
\examples{
CountNum()
}
\keyword{capitalize}
\keyword{string,}
