% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExplorer.R
\name{DataExplorer}
\alias{DataExplorer}
\title{Data Explorer}
\usage{
DataExplorer(dataset)
}
\arguments{
\item{dataset}{Dataset to investigate}
}
\description{
This function allows you to investigate a new dataset with simple visualizations. .
}
\examples{
DataExplorer()
}
\keyword{data,}
\keyword{shiny,}
\keyword{visualization}
