% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataVisualiser.R
\name{DataVisualiser}
\alias{DataVisualiser}
\title{Data Visualiser}
\usage{
DataVisualiser(dataset)
}
\arguments{
\item{dataset}{Dataset to investigate}
}
\description{
This function allows you to investigate a new dataset with simple visualizations.
}
\examples{
DataVisualiser()
}
\keyword{data,}
\keyword{shiny,}
\keyword{visualization}
