% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrettyNum.R
\name{MakePrettyNumber}
\alias{MakePrettyNumber}
\title{MakePrettyNumber}
\usage{
MakePrettyNumber(x)
}
\arguments{
\item{dataset}{Dataset to investigate}
}
\description{
This function chooses number of decimal places to display.
}
\examples{
MakePrettyNumber()
}
\keyword{capitalize}
\keyword{string,}
